pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tz"
version = "0.1.0"
description = "ToDo List System with Django, Celery, and Telegram Bot"
requires-python = ">=3.9"
dependencies = [
    "Django>=4.2.0",
    "djangorestframework>=3.14.0",
    "celery>=5.3.0",
    "django-celery-beat>=2.5.0",
    "redis>=4.6.0",
    "psycopg2-binary>=2.9.0",
    "aiogram>=3.0.0",
    "aiogram-dialog>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "django-cors-headers>=4.3.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "flake8",
    "isort",
]

[tool.setuptools.packages.find]
where = ["src"]
